requirements.txt
crewai>=1.0.0
streamlit>=1.28.0
langchain>=0.1.0
langchain_openai>=0.0.6
langchain_community>=0.0.20
voyageai>=0.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.0.0
